+++ extras/AFC.py
@@ -637,36 +637,49 @@ class afc:
 
         return wait
 
-    def _capture_toolhead_temp(self):
+    def capture_toolhead_temp(self, extruder: Optional[AFCExtruder]=None,
+                              async_capture: bool=False) -> Optional[dict]:
         """
         Helper function to capture current toolhead target temperature when not printing.
+
+        :param extruder: Pass in extruder to capture current toolhead temperature for, if no extruder
+            is passed in, defaults to current active extruder.
+        :param async_capture: Set to True to capture temp while printing, this is useful for capturing
+            other toolhead hotends on toolchangers.
 
         :return dict with extruder and target_temp, or None if printing or restore_extruder_temp_on_load_or_unload is False
         """
         if not self.restore_extruder_temp_on_load_or_unload:
             return None
-        if self.function.is_printing():
+
+        if (self.function.is_printing()
+            and not async_capture):
             return None
-        extruder = self.toolhead.get_extruder()
+
+        if extruder is None:
+            extruder = self.toolhead.get_extruder()
         heater = extruder.get_heater()
         return {"extruder": extruder, "target_temp": heater.target_temp}
 
-    def _restore_toolhead_temp(self, temp_state):
+    def restore_toolhead_temp(self, temp_state:dict, async_restore: bool=False) -> None:
         """
         Helper function to restore toolhead target temperature after load/unload when not printing AND restore_extruder_temp_on_load_or_unload is True
 
         :param temp_state: Dictionary containing extruder object and target_temp, or None
+        :param async_restore: Set to True to restore while printing, this is useful for restoring
+            other toolhead hotends on toolchangers.
         """
         if not self.restore_extruder_temp_on_load_or_unload:
             return
-        if not temp_state:
+
+        if (self.function.is_printing()
+            and not async_restore):
             return
-        if self.function.is_printing():
-            return
+
         try:
             pheaters = self.printer.lookup_object('heaters')
             pheaters.set_temperature(temp_state["extruder"].get_heater(), temp_state["target_temp"], wait=False)
-            self.logger.info("Restoring extruder temperature to {}".format(temp_state["target_temp"]))
+            self.logger.info(f"Restoring extruder temperature to {temp_state['target_temp']} for {temp_state['extruder'].name}")
         except Exception:
             self.logger.debug("Unable to restore extruder temperature", exc_info=True)
 
@@ -1184,7 +1197,7 @@ class afc:
 
         # TODO: add a check for multi-tools to verify lane is not loaded to toolhead before trying to unload
         if (cur_lane.name != cur_lane.extruder_obj.lane_loaded
-		    and not cur_lane.extruder_obj.no_lanes
+		    and not cur_lane.extruder_obj.is_standalone()
 			and not cur_lane.is_direct_hub()):
             # Setting status as ejecting so if filament is removed and de-activates the prep sensor while
             # extruder motors are still running it does not trigger infinite spool or pause logic
@@ -1206,7 +1219,7 @@ class afc:
             self.spool.set_spoolID(cur_lane, None)
             self.logger.info("LANE {} eject done".format(cur_lane.name))
             cur_lane.unit_obj.lane_not_ready(cur_lane)
-        elif cur_lane.extruder_obj.no_lanes and cur_lane.extruder_obj.lane_loaded:
+        elif cur_lane.extruder_obj.is_standalone() and cur_lane.extruder_obj.lane_loaded:
             cur_lane.status = AFCLaneState.EJECTING
             cur_lane.extruder_obj.load_unload_sequence(cur_lane.extruder_obj.tool_stn_unload*-1)
 
@@ -1328,7 +1341,7 @@ class afc:
                 self.save_vars()
                 cur_lane.unit_obj.lane_loading( cur_lane )
 
-                temp_state = self._capture_toolhead_temp()
+                temp_state = self.capture_toolhead_temp()
                 try:
                     # Run the load sequence, which may include custom gcode commands.
                     success = self.load_sequence(cur_lane, cur_hub, cur_extruder)
@@ -1383,7 +1396,7 @@ class afc:
                         self.gcode.run_script_from_command(self.post_load_macro)
                         # TODO: Add afcDeltaTime log
                 finally:
-                    self._restore_toolhead_temp(temp_state)
+                    self.restore_toolhead_temp(temp_state)
 
             else:
                 # Handle errors if the hub is not clear or the lane is not ready for loading.
@@ -1726,14 +1739,14 @@ class afc:
             # Lookup current hub object using the lane's information.
             cur_hub = cur_lane.hub_obj
 
-            temp_state = self._capture_toolhead_temp()
+            temp_state = self.capture_toolhead_temp()
             try:
                 # Run the unload sequence, which may include custom gcode commands.
                 success = self.unload_sequence(cur_lane, cur_hub, cur_extruder)
                 if not success:
                     return success
             finally:
-                self._restore_toolhead_temp(temp_state)
+                self.restore_toolhead_temp(temp_state)
 
             unload_time = self.afcDeltaTime.log_major_delta("Lane {} unload done".format(cur_lane.name if cur_lane is not None else "None"))
             self.afc_stats.average_tool_unload_time.average_time(unload_time)
@@ -1886,6 +1899,10 @@ class afc:
                         self.move_e_pos( cur_extruder.tool_stn_unload * -1, cur_extruder.tool_unload_speed, "Sensor move", wait_tool=True)
 
                     self.function.log_toolhead_pos("Sensor move after ")
+                    # For "standalone" toolheads, break out of the loop since sensor will always
+                    # be triggered
+                    if cur_lane.extruder_obj.is_standalone():
+                        break
 
             self.afcDeltaTime.log_with_time("Unloaded from toolhead")
 
@@ -1991,7 +2008,8 @@ class afc:
             cur_lane.unit_obj.lane_tool_unloaded(cur_lane)
             cur_lane.status = AFCLaneState.NONE
 
-            if cur_lane.is_direct_hub():
+            if (cur_lane.is_direct_hub()
+                and not cur_lane.extruder_obj.is_standalone()):
                 while cur_lane.raw_load_state:
                     cur_lane.move_advanced(cur_lane.short_move_dis * -1, SpeedMode.SHORT,
                                            assist_active=AssistActive.YES)