+++ extras/AFC_Toolchanger.py
@@ -19,8 +19,9 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from configfile import ConfigWrapper
     from gcode import GCodeCommand
-    from extras.AFC_lane import AFCLane
+    from extras.AFC_lane import AFCLane, MoveDirection, AFCHomingPoints
     from extras.AFC_functions import afcFunction
+    from extras.AFC_stepper import AFCExtruderStepper
 
 try: from extras.AFC_utils import ERROR_STR
 except: raise error("Error when trying to import AFC_utils.ERROR_STR\n{trace}".format(trace=traceback.format_exc()))
@@ -30,6 +31,10 @@ except: raise error(ERROR_STR.format(import_lib="AFC_unit", trace=traceback.form
 
 try: from extras.AFC import State
 except: raise error(ERROR_STR.format(import_lib="AFC", trace=traceback.format_exc()))
+
+try: from extras.AFC_lane import AFCMoveWarning, SpeedMode, AssistActive
+except: raise error(ERROR_STR.format(import_lib="AFC_lane", trace=traceback.format_exc()))
+
 
 class AfcToolchanger(afcUnit):
     def __init__(self, config: ConfigWrapper) -> None:
@@ -69,6 +74,39 @@ class AfcToolchanger(afcUnit):
         current_extruder = self.afc.function.get_current_extruder_obj()
         if current_extruder:
             current_extruder.estats.tool_unselected.increase_count()
+
+    def move_to_hub(self, lane: AFCLane, dist: float,
+                    dir: MoveDirection, use_homing: bool=True,
+                    speed_mode: SpeedMode=SpeedMode.HUB,
+                    assist_active: AssistActive=AssistActive.DYNAMIC
+                ) -> tuple[bool, float|int, AFCMoveWarning]:
+        """
+        Overriding method from AFC_unit
+
+        return: True, 0, AFCMoveWarning.NONE
+        """
+        return True, 0, AFCMoveWarning.NONE
+
+    def move_to_load(self, lane: AFCLane, dist: float,
+                     dir: MoveDirection, use_homing: bool=True,
+                     speed_mode: SpeedMode=SpeedMode.LONG
+                ) -> tuple[bool, float|int, AFCMoveWarning]:
+        """
+        Overriding method from AFC_unit
+
+        return: True, 0, AFCMoveWarning.NONE
+        """
+        return True, 0, AFCMoveWarning.NONE
+
+    def load_then_home(self, lane: AFCLane|AFCExtruderStepper, distance: float,
+                       assist_active: AssistActive, endstop: AFCHomingPoints
+                    ) -> tuple[bool, float|int, AFCMoveWarning]:
+        """
+        Overriding method from AFC_unit
+
+        return: True, 0, AFCMoveWarning.NONE
+        """
+        return True, 0, AFCMoveWarning.NONE
 
     cmd_AFC_SELECT_TOOL_help = "Select specified tool"
     cmd_AFC_SELECT_TOOL_options = {